requirements.txt
streamlit
ultralytics
langchain
groq
opencv-python-headless
Pillow
python-dotenv